TINY-11108: Moved conversion code, minor fixups.
